added last activity hash comparing to a new one
